requirements.txt
# Whereas install_requires requirements are minimal, requirements files often contain an
# exhaustive listing of pinned versions for the purpose of achieving repeatable
# installations of a complete environment.
fastapi
uvicorn
motor
pyjwt
orjson
cryptography